worker/requirements.txt
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
tenacity==9.0.0
playwright
